+++ src/components/what-we-teach/index.js
@@ -1,4 +1,8 @@
 import React from 'react';
+import tabletIcon from '../../assets/images/icons/tablet.svg';
+import codeIcon from '../../assets/images/icons/code.svg';
+import nodeJsIcon from '../../assets/images/icons/nodejs.svg';
+import reactIcon from '../../assets/images/icons/react.svg';
 
 const WhatWeTeach = () => (
   <div className="section-container section-container-gray">
@@ -15,7 +19,7 @@ const WhatWeTeach = () => (
       <div className="row">
         <div className="col-sm-3">
           <div className="md-icon">
-            <img src="img/ico/tablet.svg" alt="Tablet" />
+            <img src={tabletIcon} alt="Tablet" />
           </div>
           <h4>HTML &amp; CSS</h4>
           <p>
@@ -25,7 +29,7 @@ const WhatWeTeach = () => (
         </div>
         <div className="col-sm-3">
           <div className="md-icon">
-            <img src="img/ico/code.svg" alt="Code" />
+            <img src={codeIcon} alt="Code" />
           </div>
           <h4>JavaScript</h4>
           <p>
@@ -34,14 +38,14 @@ const WhatWeTeach = () => (
         </div>
         <div className="col-sm-3">
           <div className="md-icon">
-            <img src="img/ico/nodejs.svg" alt="NodeJS" />
+            <img src={nodeJsIcon} alt="NodeJS" />
           </div>
           <h4>Node.js</h4>
           <p>Web application development powered by modern and fast Node.js</p>
         </div>
         <div className="col-sm-3">
           <div className="md-icon">
-            <img src="img/ico/react.svg" alt="React" />
+            <img src={reactIcon} alt="React" />
           </div>
           <h4>React</h4>
           <p>A Javascript library for building user interfaces</p>
